updated insertEmployee to return int as opposed to long, to keep it consistent with the Employee object class

+++ src/main/java/dao/EmployeeDAO.java
@@ -20,5 +20,6 @@ public interface EmployeeDAO {
 			+ "employee(first_name, middle_name, last_name, address_line, post_code, email, nin, bank_sort_code, bank_account_no, salary) "
 			+ "VALUES (:firstName, :middleName, :lastName, :addressLine, :post_Code, :email, :nin, :bankSortCode, :bankAccountNo, :salary")
 	@GetGeneratedKeys
-	long insertEmployee(@BindBean Employee employee);
+	int insertEmployee(@BindBean Employee employee);
+	// the above used to return long and it worked, so change to that if there are errors
 }
